advanced_agent/pyproject.toml
[project]
name = "advanced-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "firecrawl-py>=2.8.0",
    "langchain>=0.3.25",
    "langchain-openai>=0.3.23",
    "langgraph>=0.4.8",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
]
